adjusted layouts and drawables for scrollable views

+++ Abreath/app/src/main/java/com/coen390/abreath/common/Utility.java
@@ -1,5 +1,7 @@
 package com.coen390.abreath.common;
 
+import android.util.Log;
+
 public class Utility {
     public static float map(float x, float min, float max, float out_min, float out_max){
         return (x - min) * (out_max - out_min)/(max - min) + out_min;
@@ -16,7 +18,8 @@ public class Utility {
 
     public static int[] cmtoin(float cm){
         float length = 100*(cm/100)/2.54f;
-        float feet = length / 12;
-        return new int[]{(int)feet, (int) (length - 12*feet)};
+        int feet = (int) length / 12;
+
+        return new int[]{feet, (int) (length - 12*feet)};
     }
 }
